moved the default timestamp tests to *AttributesTest
