+++ src/test/java/teammates/test/cases/common/FeedbackQuestionAttributesTest.java
@@ -3,6 +3,7 @@ package teammates.test.cases.common;
 import static teammates.common.util.Const.EOL;
 
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 
 import org.testng.annotations.AfterClass;
@@ -21,9 +22,41 @@ import com.google.appengine.api.datastore.Text;
 
 public class FeedbackQuestionAttributesTest extends BaseTestCase {
 
+    private class FeedbackQuestionAttributesWithModifiableTimestamp extends FeedbackQuestionAttributes {
+        
+        private void setCreatedAt(Date createdAt) {
+            this.createdAt = createdAt;
+        }
+        
+        private void setUpdatedAt(Date updatedAt) {
+            this.updatedAt = updatedAt;
+        }
+        
+    }
+    
     @BeforeClass
     public static void classSetUp() {
         printTestClassHeader();
+    }
+
+    @Test
+    public void testDefaultTimestamp() {
+        
+        FeedbackQuestionAttributesWithModifiableTimestamp fq =
+                new FeedbackQuestionAttributesWithModifiableTimestamp();
+        
+        fq.setCreatedAt(null);
+        fq.setUpdatedAt(null);
+        
+        Date defaultTimeStamp = Const.TIME_REPRESENTS_DEFAULT_TIMESTAMP;
+        
+        ______TS("success : defaultTimeStamp for createdAt date");
+
+        assertEquals(defaultTimeStamp, fq.getCreatedAt());
+
+        ______TS("success : defaultTimeStamp for updatedAt date");
+
+        assertEquals(defaultTimeStamp, fq.getUpdatedAt());
     }
 
     @Test
