+++ src/test/java/teammates/test/cases/storage/StudentsDbTest.java
@@ -3,8 +3,6 @@ package teammates.test.cases.storage;
 import static teammates.common.util.FieldValidator.COURSE_ID_ERROR_MESSAGE;
 import static teammates.common.util.FieldValidator.REASON_INCORRECT_FORMAT;
 
-import java.util.Date;
-
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
@@ -22,40 +20,9 @@ public class StudentsDbTest extends BaseComponentTestCase {
     
     private StudentsDb studentsDb = new StudentsDb();
     
-    private class StudentAttributesWithModifiableTimestamp extends StudentAttributes {
-        
-        private void setCreatedAt(Date createdAt) {
-            this.createdAt = createdAt;
-        }
-        
-        private void setUpdatedAt(Date updatedAt) {
-            this.updatedAt = updatedAt;
-        }
-        
-    }
-    
     @BeforeClass
     public static void setupClass() {
         printTestClassHeader();
-    }
-    
-    @Test
-    public void testDefaultTimestamp() throws InvalidParametersException {
-        
-        StudentAttributesWithModifiableTimestamp s = new StudentAttributesWithModifiableTimestamp();
-        
-        s.setCreatedAt(null);
-        s.setUpdatedAt(null);
-        
-        Date defaultStudentCreationTimeStamp = Const.TIME_REPRESENTS_DEFAULT_TIMESTAMP;
-        
-        ______TS("success : defaultTimeStamp for createdAt date");
-        
-        assertEquals(defaultStudentCreationTimeStamp, s.getCreatedAt());
-        
-        ______TS("success : defaultTimeStamp for updatedAt date");
-        
-        assertEquals(defaultStudentCreationTimeStamp, s.getUpdatedAt());
     }
     
     @Test
